requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
Django==3.1.1
django-allauth==0.44.0
django-cors-headers==3.7.0
django-emoji==2.2.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
iniconfig==1.1.1
msgpack==1.0.2
mysql==0.0.2
mysqlclient==2.0.3
oauthlib==3.1.0
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.3
pytest-asyncio==0.15.1
pytest-django==4.2.0
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.4
whitenoise==5.2.0
zope.interface==5.3.0
